Created ISwipeViewController Moved SwipeItemView to a new class Moved SwipeEventArgs to a new class Changes in touch events in Android to have the same behavior as in Android
