+++ Xamarin.Forms.Core/ISwipeViewController.cs
@@ -1,0 +1,9 @@
+﻿namespace Xamarin.Forms
+{
+	public interface ISwipeViewController
+	{
+		void SendSwipeStarted(SwipeStartedEventArgs args);
+		void SendSwipeChanging(SwipeChangingEventArgs args);
+		void SendSwipeEnded(SwipeEndedEventArgs args);
+	}
+}
