+++ Xamarin.Forms.Core/SwipeItemView.cs
@@ -1,0 +1,36 @@
+﻿using System;
+using System.ComponentModel;
+using System.Windows.Input;
+
+namespace Xamarin.Forms
+{
+	public class SwipeItemView : ContentView, ISwipeItem
+	{
+		public static readonly BindableProperty CommandProperty = BindableProperty.Create(nameof(Command), typeof(ICommand), typeof(SwipeItemView));
+
+		public static readonly BindableProperty CommandParameterProperty = BindableProperty.Create(nameof(CommandParameter), typeof(object), typeof(SwipeItemView));
+
+		public ICommand Command
+		{
+			get => (ICommand)GetValue(CommandProperty);
+			set => SetValue(CommandProperty, value);
+		}
+
+		public object CommandParameter
+		{
+			get => GetValue(CommandParameterProperty);
+			set => SetValue(CommandParameterProperty, value);
+		}
+
+		public event EventHandler<EventArgs> Invoked;
+
+		[EditorBrowsable(EditorBrowsableState.Never)]
+		public void OnInvoked()
+		{
+			if (Command != null && Command.CanExecute(CommandParameter))
+				Command.Execute(CommandParameter);
+
+			Invoked?.Invoke(this, EventArgs.Empty);
+		}
+	}
+}
